refactor(daemon): split ReviewBotRepoManager into persistent ReviewBranchManager ops
Split the one-shot PublishAsync (create branch, commit, ff-merge to
default, push, delete branch) into three independently-callable ops
so the PR review-notes branch can persist across re-reviews and be
merged-or-deleted only when the PR closes:

- CommitNotesAsync: create-or-reuse the review branch, write notes,
  commit, push the branch. Never merges or deletes.
- MergeToDefaultAsync: merge the branch into default (ff-only, else a
  merge commit), push default, then delete the branch. Returns false
  and keeps the branch if the push never succeeds.
- DeleteBranchAsync: idempotent branch delete (local + remote).

DaemonReviewStageExecutor now calls CommitNotesAsync in place of
PublishAsync; wiring MergeToDefaultAsync/DeleteBranchAsync into the
PR-close lifecycle is a follow-up task. Updated the executor tests
that previously asserted a push to `main` to assert the review branch
is committed, pushed, and kept (not merged or deleted).
